requirements.txt
# Agent Project Requirements

# --- Core Framework ---
aiogram>=3.4.0 # Используем актуальную версию aiogram (на момент создания)
pydantic>=2.0.0
pydantic-settings>=2.0.0 # Для удобной загрузки .env в Pydantic

# --- AI Interface ---
google-generativeai>=0.5.0 # Актуальная версия Gemini API
google-api-python-client # Зависимость для google-generativeai
google-auth # Зависимость для google-generativeai
protobuf # Зависимость для google-generativeai

# --- Database ---
aiosqlite>=0.17.0 # Асинхронный драйвер SQLite
# sqlalchemy>=2.0.0 # Раскомментировать, если будете использовать SQLAlchemy ORM
# alembic>=1.10.0 # Раскомментировать для миграций с SQLAlchemy
# psycopg2-binary # Раскомментировать, если будете использовать PostgreSQL

# --- Tools & Services Dependencies ---
aiohttp>=3.9.0 # Асинхронные HTTP запросы (для аватаров, парсинга чартов, RSS)
aiofiles>=23.0.0 # Асинхронные операции с файлами
beautifulsoup4>=4.11.0 # Парсинг HTML (для чартов, RSS)
lxml>=4.9.0 # Быстрый парсер для BeautifulSoup (опционально, но рекомендуется)
requests>=2.30.0 # Синхронные HTTP запросы (может требоваться для некоторых старых частей парсеров)
feedparser>=6.0.0 # Парсинг RSS лент
duckduckgo-search>=5.0.0 # Для инструмента веб-поиска

# --- Environment & Utilities ---
python-dotenv>=1.0.0 # Загрузка .env файлов

# --- Development & Optional ---
# pytest # Для запуска тестов
# pytest-asyncio # Для асинхронных тестов
# flake8 # Линтер
# mypy # Статический анализатор типов
# pre-commit # Для git hooks